requirements.txt
allure-pytest==2.8.12
mysql-connector-python==8.0.19
pytest==5.4.1
requests==2.23.0
selenium==3.141.0